+++ lib/index_generator.dart
@@ -1,7 +1,7 @@
 library index_generator;
 
+export 'src/converters.dart';
+export 'src/converters.dart';
 export 'src/index_generator.dart';
-export 'src/settings/filters.dart';
-export 'src/settings/filters.dart';
 export 'src/settings/package_settings.dart';
 export 'src/stringify.dart';